Split smart-status conflict tests
